Fixed adding Vehicles & Customers

+++ Car Rental.Common/Classes/Booking.cs
@@ -17,16 +17,16 @@ public class Booking : IBooking
 
     public DateTime Start { get; init; }
     public DateTime? End { get; init; }
-    public int Cost { get; init; }
+    public double Cost { get; init; }
 
-    public int KmRented { get; init; }
-    public int KmReturned { get; init; }
+    public double KmRented { get; init; }
+    public double KmReturned { get; init; }
     public DateTime days { get; init; }
 
     public VehicleStatuses VStatus { get; init; }
  
     
-    public Booking(int id, ICustomer customer, IVehicle vehicle,   /*int customerSSN,  string customer, string regNo, VehiclesMake vehiclesMake,*/ int cost, /*int kmRented,*/ int kmReturned, DateTime start, DateTime end, VehicleStatuses vStatus)
+    public Booking(int id, ICustomer customer, IVehicle vehicle,   /*int customerSSN,  string customer, string regNo, VehiclesMake vehiclesMake,*/ double cost, /*int kmRented,*/ double kmReturned, DateTime start, DateTime end, VehicleStatuses vStatus)
    => (Id, Person, Vehicle, /*CustomerSSN, Customer, RegNo, VehiclesMake,*/ Cost, /* KmRented, */ KmReturned, Start, End, VStatus)
     = (id, customer, vehicle, /*customerSSN, customer, regNo, vehiclesMake,*/ cost, /* kmRented, */  kmReturned, start, end, VStatus);
