+++ Car Rental.Common/Classes/Motorcycle.cs
@@ -13,10 +13,10 @@ public class Motorcycle : Vehicle
     int McID {  get; set; }
 
     #region Constructors
-    public Motorcycle(int id, string regNo, VehiclesMake vMake, VehiclesTypes vType, int odometer, int costKm, int costDay, VehicleStatuses vStatus)
+    public Motorcycle(int id, string regNo, VehiclesMake vMake, VehiclesTypes vType, double odometer, double costKm, double costDay, VehicleStatuses vStatus)
         : base( id, regNo, vMake, vType, odometer, costKm, costDay, vStatus ) { }
 
-    public Motorcycle( int mcid, int id, string regNo, VehiclesMake vMake, VehiclesTypes vType, int odometer, int costKm, int costDay, VehicleStatuses vStatus)
+    public Motorcycle( int mcid, int id, string regNo, VehiclesMake vMake, VehiclesTypes vType, double odometer, double costKm, double costDay, VehicleStatuses vStatus)
       : base(id, regNo, vMake, vType, odometer, costKm, costDay, vStatus) { McID = mcid; }
 
     #endregion
