+++ Car Rental.Common/Extensions/BookingExtensions.cs
@@ -9,14 +9,14 @@ public static class BookingExtensions
     public static (DateTime, DateTime) BookingDates(DateTime startingDate, DateTime endingDate)
     => (startingDate, endingDate);
 
-    public static int CostCalculator(IVehicle vehicle, DateTime startBook, DateTime endBook, int kmReturned)
+    public static double CostCalculator(IVehicle vehicle, DateTime startBook, DateTime endBook, double kmReturned)
     {
 
         TimeSpan numberOfDays = endBook - startBook;
 
-        int totalDays = (int)numberOfDays.TotalDays;
+        double totalDays = (double)numberOfDays.TotalDays;
 
-        int cost = 0;
+        double cost = 0;
 
         if (endBook == DateTime.MinValue)
         {
