+++ Car Rental.Common/Interfaces/IBooking.cs
@@ -18,9 +18,9 @@ namespace Car_Rental.Common.Interfaces
 
         public DateTime Start { get; init; }
         public DateTime? End { get; init; }
-        public int Cost { get; init; }
+        public double Cost { get; init; }
        // public int KmRented { get; init; }
-        public int KmReturned { get; init; }
+        public double KmReturned { get; init; }
         public DateTime days { get; init; }
 
        // public VehicleStatuses VStatus { get; init; }
